requirements.txt
# Web Framework
streamlit>=1.28.0

# Google Gemini API
google-generativeai>=0.8.0

# LangChain Core & Extensions
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langchain-experimental>=0.0.50
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.2.0
langchain-classic>=0.1.0

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Utilities
pyyaml>=6.0